// packages/core/generated/types.ts
export interface StructField {
  offset: number;
  ffi: Deno.NativeType;
  count: number;
  cType: string;
}
export interface StructLayout {
  size: number;
  align: number;
  fields: Record<string, StructField>;
}